updates NoteAPITest.kt
I have added functions to test the load function and the store function.

I have encountered an error in the NoteAPITest.kt but I cannot fix it

I will work on the JSON persistence now.
